Refactor BzPath handling and update interpolation methods
- Updated the `BzPath` class to improve clarity and organization in the state space module.
- Refactored the `interpolate_path` function to alias it as `interpolate_reciprocal_path`, enhancing consistency in naming conventions.
- Adjusted import statements across multiple files for better structure and readability.
- Updated documentation and comments to reflect changes in method names and functionality.
- Enhanced test cases to validate the new interpolation method and ensure compatibility with existing functionality.
