+++ src/qten/geometries/__init__.py
@@ -12,4 +12,8 @@ from .boundary import (
 )
 
 from .basis_transform import BasisTransform as BasisTransform
-from .ops import nearest_sites as nearest_sites
+from .ops import (
+    nearest_sites as nearest_sites,
+    interpolate_reciprocal_path as interpolate_reciprocal_path,
+    interpolate_path as interpolate_path,
+)
