+++ src/qten/ops.py
@@ -2,7 +2,10 @@ from .geometries.fourier import (
     fourier_transform as fourier_transform,
     region_restrict as region_restrict,
 )
-from .geometries import nearest_sites as nearest_sites
+from .geometries import (
+    nearest_sites as nearest_sites,
+    interpolate_reciprocal_path as interpolate_reciprocal_path,
+)
 from .symbolics import (
     region_hilbert as region_hilbert,
     hilbert_opr_repr as hilbert_opr_repr,
