+++ tests/test_plots.py
@@ -7,7 +7,7 @@ import pytest
 import sympy as sy
 import torch
 
-from qten.bands import interpolate_path
+from qten.geometries import interpolate_reciprocal_path
 from qten.symbolics.hilbert_space import U1Basis, HilbertSpace
 from qten.geometries.spatials import Lattice, Offset
 from qten.geometries.boundary import PeriodicBoundary
@@ -423,7 +423,7 @@ def _make_2d_lattice_and_spaces():
 
 def test_bandstructure_plot_with_bz_path_plotly():
     lat, bloch_space, region_space, h_real = _make_2d_lattice_and_spaces()
-    path = interpolate_path(
+    path = interpolate_reciprocal_path(
         lat.dual,
         [(0, 0), (0.5, 0), (0.5, 0.5), (0, 0)],
         n_points=30,
@@ -456,7 +456,7 @@ def test_bandstructure_plot_with_bz_path_matplotlib():
     import matplotlib.pyplot as plt
 
     lat, bloch_space, region_space, h_real = _make_2d_lattice_and_spaces()
-    path = interpolate_path(
+    path = interpolate_reciprocal_path(
         lat.dual,
         [(0, 0), (0.5, 0), (0.5, 0.5), (0, 0)],
         n_points=30,
